Refactor build.py to improve build file cleanup process and enhance cross-platform compatibility

+++ build.py
@@ -2,13 +2,23 @@ from PyInstaller.__main__ import run
 import os
 
 if __name__ == "__main__":
-    os.system("rm -rf build dist SaveFileControl.spec")
-    
+    # delete previous build files (windows, mac and linux support)
+    if os.path.exists("dist"):
+        os.system("rmdir /s /q dist" if os.name == "nt" else "rm -rf dist")
+    if os.path.exists("build"):
+        os.system("rmdir /s /q build" if os.name == "nt" else "rm -rf build")
+    if os.path.exists("SaveFileControl.spec"):
+        os.remove("SaveFileControl.spec")
+
     run([
-        'main.py',
-        '--noconsole',
-        '--onefile',
-        '--name', 'SaveFileControl'
+        "main.py",
+        "--noconsole",
+        "--onefile",
+        "--name", "SaveFileControl"
     ])
     
-    os.system("rm -rf build SaveFileControl.spec")
+    # clean up build files
+    if os.path.exists("build"):
+        os.system("rmdir /s /q build" if os.name == "nt" else "rm -rf build")
+    if os.path.exists("SaveFileControl.spec"):  
+        os.system("rm -rf build SaveFileControl.spec")
